+++ src/Anki/TestModePanel.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Card, QuestionData, TransformedDataAndSetter, getTimeUntilNextTest } from "./mvc/model";
+import { TransformedDataAndSetter, flattenItems } from "./mvc/model";
 import { Button } from '@mui/material';
 
 const RETEST_IMMINENTLY_TIME = 60 * 60 * 1000;
@@ -11,36 +11,27 @@ export const TestModePanel = (props: {
     const testableItems = Object.entries(items);
     const [reveal, setReveal] = React.useState(false);
     if (testableItems.length) {
-        const itemToTestNow: [string, Card] = testableItems.reduce((prev, item) => {
-            const getItemTestability = (questionData: QuestionData): number => {
-                return -Math.min(...Object.values(questionData).map((data) => getTimeUntilNextTest(data)));
-            }
-            return getItemTestability(prev[1].questionData) > getItemTestability(item[1].questionData) ? prev : item
-        }, testableItems[0]);
-        const testCandidates = Object.entries(itemToTestNow[1].questionData);
-        const mostTestable = testCandidates.reduce((prev, next) => getTimeUntilNextTest(prev[1]) < getTimeUntilNextTest(next[1]) ? prev : next, testCandidates[0]);
-        let testText = itemToTestNow[1].data;
-
+        const allItemsByTestability = flattenItems(items);
+        allItemsByTestability.sort((a, b) => a.testability - b.testability);
+        const mostTestable = allItemsByTestability[0];
         const updateCurrentCard = (goodOrBad: "good" | "bad") => {
             props.transformedDataAndSetter.setItemsByKey((transformedData) => {
-                const itemToUpdate = transformedData[itemToTestNow[0]];
+                const itemToUpdate = transformedData[mostTestable.id];
                 const newQuestionData = { ...itemToUpdate.questionData };
                 if (goodOrBad == "good") {
-                    newQuestionData[mostTestable[0]].familiarity *= 1.1;
-                    newQuestionData[mostTestable[0]].lastTested = Date.now();
+                    newQuestionData[mostTestable.questionId].familiarity *= 1.1;
+                    newQuestionData[mostTestable.questionId].lastTested = Date.now();
                 } else {
-                    newQuestionData[mostTestable[0]].familiarity *= 0.9;
-                    newQuestionData[mostTestable[0]].lastTested = Date.now() - RETEST_IMMINENTLY_TIME;
+                    newQuestionData[mostTestable.questionId].familiarity *= 0.9;
+                    newQuestionData[mostTestable.questionId].lastTested = Date.now() - RETEST_IMMINENTLY_TIME;
                 }
-                return { [itemToTestNow[0]]: { ...itemToUpdate, questionData: newQuestionData } };
+                return { [mostTestable.id]: { ...itemToUpdate, questionData: newQuestionData } };
             })
         }
 
-        if (!reveal) testText = testText.replace(mostTestable[0], "???");
-        testText = testText.replace(/\{\{|\}\}/g, "")
         return <>
-            <h2>{testText}</h2>
-            <h3>This card repeats every: {toDurationString(mostTestable[1].familiarity)}</h3>
+            <h2>{reveal ? mostTestable.revealText : mostTestable.testText}</h2>
+            <h3>This card repeats every: {toDurationString(mostTestable.testability)}</h3>
             <div className="anki-buttons">
                 <Button fullWidth variant="contained" onClick={() => setReveal(true)}>Reveal</Button>
                 <Button fullWidth variant="contained" onClick={() => { updateCurrentCard("good"); setReveal(false) }}>Good</Button>
@@ -72,9 +63,9 @@ export const toDurationString = (millis: number): string => {
             millis -= wholePortions * durationMillis;
         }
     }
-    if (builder.length > 0){
+    if (builder.length > 0) {
         return builder.join(", ");
-    }else{
+    } else {
         return "Less than a minute";
     }
 }